% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bread_glmer.R
\name{bread.glmerMod}
\alias{bread.glmerMod}
\title{glmer bread matrices}
\usage{
bread.glmerMod(x, grad_options = NULL, ...)
}
\arguments{
\item{grad.method}{method passed to \code{\link[numDeriv]{grad}}}

\item{grad.method.args}{method.args passed to \code{\link[numDeriv]{grad}}}
}
\description{
Currently supports Logistic-Normal model with univariate random effect.
}

