% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun.geeglm.R
\name{estfun.geeglm}
\alias{estfun.geeglm}
\title{geeglm Estimating Equations}
\usage{
\method{estfun}{geeglm}(x, ...)
}
\arguments{
\item{x}{\code{geeglm} object}
}
\description{
geeglm Estimating Equations
}

