% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_cluster.R
\name{estfun.glm_cluster}
\alias{estfun.glm_cluster}
\title{Estimating Equations for clustered data}
\usage{
\method{estfun}{glm_cluster}(x, groups, ...)
}
\arguments{
\item{x}{\code{lm} or \code{glm} object}

\item{groups}{vector of groupings}
}
\description{
Estimating Equations for clustered data
}

