% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{estfun.glmerMod}
\alias{estfun.glmerMod}
\title{glmer Estimating Equations}
\usage{
\method{estfun}{glmerMod}(x, grad_method = "simple", grad_options = NULL)
}
\arguments{
\item{x}{a \link[lme4]{merMod-class} object}

\item{grad_method}{method passed to \code{\link[numDeriv]{grad}}}

\item{grad_options}{list of options passed to \code{\link[numDeriv]{grad}}
'method.opts' argument}
}
\value{
a (number of groups) x (number of parameters) matrix
}
\description{
Currently supports Logistic-Normal model with univariate random effect.
}

