% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_cluster.R
\name{estfun_cluster}
\alias{estfun_cluster}
\title{Estimating Equations for clustered data}
\usage{
estfun_cluster(x, clusters, ...)
}
\arguments{
\item{x}{\code{lm} or \code{glm} object}

\item{groups}{vector of groupings}
}
\description{
Estimating Equations for clustered data
}

