% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_stacker.R
\name{estfun_stacker}
\alias{estfun_stacker}
\title{Stacks Estimating Equations from multiple models using same data}
\usage{
estfun_stacker(models, data)
}
\arguments{
\item{models}{list of lists of models with following arguments: `formula`,
`method`, and `options`}

\item{data}{data.frame used in models}
}
\description{
Stacks Estimating Equations from multiple models using same data
}

