% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_bread.R
\name{make_bread}
\alias{make_bread}
\title{Creates a block diagonal bread matrix from multiple models}
\usage{
make_bread(models, grad_method = NULL)
}
\arguments{
\item{models}{list of lists of models with following arguments: `formula`,
`method`, and `options`}
}
\description{
Creates a block diagonal bread matrix from multiple models
}

