% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{make_models}
\alias{make_models}
\title{Makes model objects from list of models}
\usage{
make_models(model_args, data)
}
\arguments{
\item{data}{used for models}

\item{models_args}{list of model arguments}
}
\description{
Makes model objects from list of models
}

