% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun.glmer.R
\name{objFun.glmer.binomial}
\alias{objFun.glmer.binomial}
\title{glmer Objective Function for Logistic-Normal Likelihood}
\usage{
objFun.glmer.binomial(parms, response, xmatrix)
}
\arguments{
\item{response}{}

\item{xmatrix}{}

\item{vector}{of parameters}
}
\description{
glmer Objective Function for Logistic-Normal Likelihood
}

