% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{objFun.glmerMod.generalized}
\alias{objFun.glmerMod.generalized}
\title{glmer Objective Function for Likelihoods specified by glmerMod object}
\usage{
objFun.glmerMod.generalized(parms, response, xmatrix, inv_link_fun)
}
\arguments{
\item{response}{}

\item{xmatrix}{}

\item{inv_link_fun}{}

\item{vector}{of parameters}
}
\description{
glmer Objective Function for Likelihoods specified by glmerMod object
}

