% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun.glmer.R
\name{objFun.glmerMod}
\alias{objFun.glmerMod}
\title{glmer Objective Fundtion}
\usage{
objFun.glmerMod(family, ...)
}
\arguments{
\item{family}{}

\item{...}{additional arguments pass to objective function}
}
\description{
glmer Objective Fundtion
}

