% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{objFun_glmerMod_binomial}
\alias{objFun_glmerMod_binomial}
\title{glmer Objective Function for Logistic-Normal Likelihood}
\usage{
objFun_glmerMod_binomial(parms, response, xmatrix, linkinv)
}
\arguments{
\item{parms}{vector of parameters}

\item{response}{}

\item{xmatrix}{}

\item{linkinv}{inverse link function}
}
\description{
glmer Objective Function for Logistic-Normal Likelihood
}

