% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{psi.glmerMod}
\alias{psi.glmerMod}
\title{glmer Estimating Equations}
\usage{
psi.glmerMod(x, grad_options = NULL, hessian, ...)
}
\arguments{
\item{grad.method}{method passed to \code{\link[numDeriv]{grad}}}

\item{grad.method.args}{method.args passed to \code{\link[numDeriv]{grad}}}
}
\description{
Currently supports Logistic-Normal model with univariate random effect.
}

