% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bread_glmer.R
\name{bread.glmerMod}
\alias{bread.glmerMod}
\title{glmer bread matrices}
\usage{
\method{bread}{glmerMod}(x, grad_method = "Richardson", grad_options = NULL,
  ...)
}
\arguments{
\item{x}{`glmer` object}

\item{grad_method}{method passed to \code{\link[numDeriv]{hessian}}}

\item{grad_options}{method.args passed to \code{\link[numDeriv]{hessian}}}
}
\description{
Currently supports Logistic-Normal model with single random effect.
}

