% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun.glmer.R
\name{estfun.glmerMod}
\alias{estfun.glmerMod}
\title{glmer Estimating Equations}
\usage{
\method{estfun}{glmerMod}(x, grad.method, grad.method.args = list(), ...)
}
\arguments{
\item{grad.method}{method passed to \code{\link[numDeriv]{grad}}}

\item{grad.method.args}{method.args passed to \code{\link[numDeriv]{grad}}}
}
\description{
Currently supports Logistic-Normal model with univariate random effect.
}

