% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun.geeglm.R
\name{gesteqn}
\alias{gesteqn}
\title{Formulas for Generalized Estimating Equations}
\usage{
gesteqn(xmat, w, lp, f, r, psi, family_link)
}
\arguments{
\item{w}{vector of weights}

\item{f}{fitted values}

\item{r}{residuals (Y - mu)}

\item{psi}{dispersion parameter}

\item{x}{the model matrix}

\item{lc}{linear predictor (X * beta)}

\item{family}{model family}
}
\description{
Formulas for Generalized Estimating Equations
}

