% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{objFun_glmerMod}
\alias{objFun_glmerMod}
\title{glmer Objective Fundtion}
\usage{
objFun_glmerMod(family, ...)
}
\arguments{
\item{family}{}

\item{...}{additional arguments pass to objective function}
}
\description{
glmer Objective Fundtion
}

