% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estfun_glmer.R
\name{psi.glmerMod}
\alias{psi.glmerMod}
\title{glmer Estimating Equations}
\usage{
psi.glmerMod(x, grad_method, grad_options = NULL, deriv, ...)
}
\arguments{
\item{x}{`glmer` object}

\item{grad_method}{method passed to \code{\link[numDeriv]{grad}} or
\code{\link[numDeriv]{hessian}} when `hessian = TRUE`}

\item{grad_options}{method.args passed to \code{\link[numDeriv]{grad}} or
\code{\link[numDeriv]{hessian}} when `hessian = TRUE`}

\item{deriv}{logical indicating whether to create derivative matrix of psi}
}
\description{
Currently supports Logistic-Normal model with univariate random effect.
}

